Extract Form DSL into a separate module
